Give up on treating the channel axis differently during transpose
Turns out it's very hard to fully support this because metadata
that refers to specific axes in the original file needs to be
manipulated properly if the array data's axes are messed with, and doing
this in a general way is somewhere between hard and not actually
sensible to try to do. We instead can have functions that care about
where a channel axis is just check where it is and act appropriately.