Cache the frame iterator. Huge improvement in reliability!
Turns out that when we `container.seek`, it changes the pointer
such that any existing iterators created previously with
`containter.decode` are affected and will return frames at a new
location determined by the result of the seek operations. This means we
don't have to call `container.decode` each time we want to start
reading frames, we can just call it once in `__init__`. And it turns out
this avoids a bug where calling `container.decode` each time caused
skipped frames!